Added support for the new `print` and `println` keywords.
Also added properties to indicate when a keyword was introduced (`available`) and when it was deprecated (`deprecated`). Currently unused, but in the future it could be possible to let the user specify in a configuration the version of RGBDS they're using (while defaulting to assuming the latest), and some behaviors could be customized based on that version selection.