Added support for rb/rw/rl defs without values
RGBDS allows values to be omitted from `rb`, `rw`, and `rl` definitions, in which case an assumed value will be inferred (1, 2, and 4, respectively). This commit allows the value to be omitted in the define expression regex.

This commit also slightly adjusts the `equ`/`equs`/`set`/`=` case to require a value to actually be assigned. Previously, the regex would match with just whitespace after the keyword.